requirements.txt
requests
pypdf
docxtpl
python-docx
reportlab